better folder structure and some optimizations
