+++ src/components/Form/Login.tsx
@@ -6,8 +6,8 @@ import { router } from "expo-router";
 import React, { Dispatch, FC, SetStateAction } from "react";
 import { Button, Form, H2, XStack, YStack } from "tamagui";
 
+import { loginUser } from "@/api/auth";
 import { useUser } from "@/hooks/useUser";
-import { loginUser } from "@/utils/loginUser";
 
 import MySheet from "../common/MySheet";
